windows support fixes

+++ tests/runtime_test.py
@@ -3,39 +3,42 @@ import signal
 import os
 
 import stackimpact
-from stackimpact.runtime import register_signal
+from stackimpact.runtime import runtime_info, register_signal
 
 
 class RuntimeTestCase(unittest.TestCase):
 
     def test_register_signal(self):
-    	result = {'handler': 0}
+        if runtime_info.OS_WIN:
+            return
 
-    	def _handler(signum, frame):
-    		result['handler'] += 1
+        result = {'handler': 0}
 
-    	register_signal(signal.SIGUSR1, _handler)
+        def _handler(signum, frame):
+            result['handler'] += 1
 
-    	os.kill(os.getpid(), signal.SIGUSR1)
-    	os.kill(os.getpid(), signal.SIGUSR1)
+        register_signal(signal.SIGUSR1, _handler)
 
-    	signal.signal(signal.SIGUSR1, signal.SIG_DFL)
+        os.kill(os.getpid(), signal.SIGUSR1)
+        os.kill(os.getpid(), signal.SIGUSR1)
 
-    	self.assertEqual(result['handler'], 2)
+        signal.signal(signal.SIGUSR1, signal.SIG_DFL)
+
+        self.assertEqual(result['handler'], 2)
 
 
     '''def test_register_signal_default(self):
-    	result = {'handler': 0}
+        result = {'handler': 0}
 
-    	def _handler(signum, frame):
-    		result['handler'] += 1
+        def _handler(signum, frame):
+            result['handler'] += 1
 
-    	register_signal(signal.SIGUSR1, _handler, once = True)
+        register_signal(signal.SIGUSR1, _handler, once = True)
 
-    	os.kill(os.getpid(), signal.SIGUSR1)
-    	os.kill(os.getpid(), signal.SIGUSR1)
+        os.kill(os.getpid(), signal.SIGUSR1)
+        os.kill(os.getpid(), signal.SIGUSR1)
 
-    	self.assertEqual(result['handler'], 1)'''
+        self.assertEqual(result['handler'], 1)'''
 
 
 if __name__ == '__main__':
